pyproject.toml
[project]
name = "graphiti-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "graphiti-core>=0.24.1",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
]
